Remove the src argument from the decoder forward() interface
After split, forward() of StdRNNDecoder and InputFeedRNNDecoder
don't need the `src` argument, so we can easily remove it.

For TransformerDecoder, it needs `src` argument to calculate the
mask, and this actully relates to each batch, so pass this `src`
into TransformerDecoderState when calling init_decoder_state()
for each batch.

So now we can get rid of the src argument.