Trainer: simplify constructor, pass in train_loss/valid_loss
Now that we get rid of the epoch arg in LossCompute in previous commit,
we don't need to create the LossCompute instances each epoch, just create
them once ans pass in to the constructor.

LossCompute is the loss computation interface the trainer calls to compute
loss in each valid process(the mechanism), and it should be agnostic to
which loss criterion is used(the strategy).